requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
python-magic==0.4.27
sqlparse==0.5.0
typing_extensions==4.12.1
